Refactor tests. Fixes #45 and fixes #47
